refractored tests
