+++ src/test/java/edu/ntnu/idatt2003/group16/model/filemanagement/StockFileReaderTest.java
@@ -1,4 +1,4 @@
-package edu.ntnu.idatt2003.group16.filemanagement;
+package edu.ntnu.idatt2003.group16.model.filemanagement;
 
 import edu.ntnu.idatt2003.group16.model.filemanagement.StockFileReader;
 import edu.ntnu.idatt2003.group16.model.investment.Stock;
