+++ src/test/java/edu/ntnu/idatt2003/group16/model/investment/PortfolioTest.java
@@ -1,4 +1,4 @@
-package edu.ntnu.idatt2003.group16.investment;
+package edu.ntnu.idatt2003.group16.model.investment;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertFalse;
