+++ src/test/java/edu/ntnu/idatt2003/group16/model/player/PlayerTest.java
@@ -1,4 +1,4 @@
-package edu.ntnu.idatt2003.group16.player;
+package edu.ntnu.idatt2003.group16.model.player;
 
 import static org.junit.jupiter.api.Assertions.assertDoesNotThrow;
 import static org.junit.jupiter.api.Assertions.assertEquals;
